+++ tgbotapi.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/edit/LiveLocation/EditInlineMessageLiveLocation.kt
@@ -1,7 +1,7 @@
 package dev.inmo.tgbotapi.extensions.api.edit.LiveLocation
 
 import dev.inmo.tgbotapi.bot.TelegramBot
-import dev.inmo.tgbotapi.requests.edit.LiveLocation.EditInlineMessageLiveLocation
+import dev.inmo.tgbotapi.requests.edit.location.live.EditInlineMessageLiveLocation
 import dev.inmo.tgbotapi.types.*
 import dev.inmo.tgbotapi.types.buttons.InlineKeyboardMarkup
 import dev.inmo.tgbotapi.types.location.LiveLocation