+++ tgbotapi.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/edit/LiveLocation/StopInlineMessageLiveLocation.kt
@@ -1,7 +1,7 @@
 package dev.inmo.tgbotapi.extensions.api.edit.LiveLocation
 
 import dev.inmo.tgbotapi.bot.TelegramBot
-import dev.inmo.tgbotapi.requests.edit.LiveLocation.StopInlineMessageLiveLocation
+import dev.inmo.tgbotapi.requests.edit.location.live.StopInlineMessageLiveLocation
 import dev.inmo.tgbotapi.types.InlineMessageIdentifier
 import dev.inmo.tgbotapi.types.buttons.InlineKeyboardMarkup